--- cockroachdb/README.md
@@ -1,3 +1,0 @@
-# Package `cockroachdb`
-
-Useful tools around CockroachDB.
